feat: split helper functions to util file
