+++ src/main.rs
@@ -1,9 +1,10 @@
 mod args;
+mod utils;
 
 use reqwest;
 use serde_json::{self, Value};
 use tokio;
-use colored::*;
+
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -33,14 +34,14 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                 spotlight.clone().unwrap().to_uppercase().eq(away_team_tricode) ||
                 spotlight.clone().unwrap().to_uppercase().eq(home_team_tricode)
             ) {
-                display(game);
+                utils::display_per_game(game);
                 break;
             } else if spotlight.is_some() {
                 println!("😔 That team is not currently playing");
                 break;
             }
 
-            display(game);
+            utils::display_per_game(game);
         }
     }
     else{
@@ -48,28 +49,5 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     }
     
     Ok(())
-
-}
-
-fn display(game: &Value) {
-
-    let away_team = game.get("awayTeam").unwrap();
-    let home_team = game.get("homeTeam").unwrap();
-
-    let away_team_tricode = away_team.get("teamTricode").unwrap();
-    let away_team_score = away_team.get("score").unwrap();
-
-    let home_team_tricode = home_team.get("teamTricode").unwrap();
-    let home_team_score = home_team.get("score").unwrap();
-
-    let game_status = game.get("gameStatusText").unwrap();
-
-    println!("🏀 {}: {} - {}: {} ({})", 
-        away_team_tricode.as_str().unwrap().bold(),
-        away_team_score.to_string().bright_blue(),
-        home_team_tricode.as_str().unwrap().bold(),
-        home_team_score.to_string().bright_blue(),
-        game_status.as_str().unwrap().trim().yellow()
-    );
 
 }
